requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0

# Dashboard
streamlit>=1.29.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
